refactor: put action functions to separate modules
